feat: add read process cpu and memory

+++ src/platform/macos.rs
@@ -93,7 +93,7 @@ impl Measurement for MeasurementImpl {
         Err(io::Error::new(io::ErrorKind::Other, "Not supported"))
     }
 
-    fn memory_by_pid(&self, _pid: u32) -> std::io::Result<(f64, f64)> {
+    fn memory_by_pid(&self, _pid: u32) -> std::io::Result<(u64, u64)> {
         Err(io::Error::new(io::ErrorKind::Other, "Not supported"))
     }
